Enable native hooks lib to be statically linked
